+++ AzureCommunicationUI/sdk/AzureCommunicationUICalling/Sources/Presentation/FluentUI/Wrapper/DrawerContainerViewController.swift
@@ -6,7 +6,9 @@
 import FluentUI
 import UIKit
 
-class DrawerContainerViewController<T: Equatable>: UIViewController, DrawerControllerDelegate {
+class DrawerContainerViewController<T: Equatable>: UIViewController,
+                                                    DrawerControllerDelegate,
+                                                    DrawerViewControllerProtocol {
     weak var delegate: DrawerControllerDelegate?
     lazy var drawerTableView: UITableView? = nil
     let backgroundColor: UIColor = UIDevice.current.userInterfaceIdiom == .pad
@@ -54,15 +56,11 @@ class DrawerContainerViewController<T: Equatable>: UIViewController, DrawerContr
 
     override func viewDidDisappear(_ animated: Bool) {
         super.viewDidDisappear(animated)
-        print("!!!! drawer viewDidDisappear")
         if isBeingDismissed || isMovingFromParent {
             sourceView.superview?.isUserInteractionEnabled = true
             sourceView.removeFromSuperview()
         }
-        if UIDevice.current.userInterfaceIdiom == .phone {
-            UIDevice.current.setValue(UIDevice.current.orientation.rawValue, forKey: "orientation")
-            UIViewController.attemptRotationToDeviceOrientation()
-        }
+        resetOrientation()
     }
 
     override func viewWillTransition(to size: CGSize, with coordinator: UIViewControllerTransitionCoordinator) {
